Fix small error in iOS documentation
This fixes a small error in the iOS documentation that I ran into. The old documentation stated that the bridge components must be registered before routing your first URL. However, because the userAgent is checked to determine if we are in a Hotwire app and therefore if the bridge components should be activated **and** because this userAgent is established when the root view controller is created you need to register your bridge components first.