Clarify defaultViewController and defaultNavigationController
The current documentation states the protocol requirements for the `defaultViewController` and `defaultNavigationController` (`VisitableViewController` and `UINavigationController` respectively), but doesn't mention `HotwireWebViewController` or `HotwireNavigationController`. As such, developers may just subclass `VisitableViewController` for example, only to discover that their Bridge Components don't work.

This PR informs developers of these default classes so that they can create the appropriate subclass.